requirements.txt
Django==1.9.5
Pillow==3.3.1
argparse==1.2.1
django-registration-redux==1.4
wsgiref==0.1.2
